+++ adetailer/common.py
@@ -1,13 +1,15 @@
 from __future__ import annotations
 
+import os
 from collections import OrderedDict
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Optional
+from typing import Any, Optional
 
 from huggingface_hub import hf_hub_download
 from PIL import Image, ImageDraw
 from rich import print
+from torchvision.transforms.functional import to_pil_image
 
 REPO_ID = "Bingsu/adetailer"
 _download_failed = False
@@ -36,16 +38,27 @@ def hf_download(file: str, repo_id: str = REPO_ID) -> str | None:
     return path
 
 
-def scan_model_dir(path_: str | Path) -> list[Path]:
-    if not path_ or not (path := Path(path_)).is_dir():
+def safe_mkdir(path: str | os.PathLike[str]) -> None:
+    path = Path(path)
+    if not path.exists() and path.parent.exists() and os.access(path.parent, os.W_OK):
+        path.mkdir()
+
+
+def scan_model_dir(path: Path) -> list[Path]:
+    if not path.is_dir():
         return []
-    return [p for p in path.rglob("*") if p.is_file() and p.suffix in (".pt", ".pth")]
+    return [p for p in path.rglob("*") if p.is_file() and p.suffix == ".pt"]
 
 
 def get_models(
-    model_dir: str | Path, extra_dir: str | Path = "", huggingface: bool = True
-) -> OrderedDict[str, str | None]:
-    model_paths = [*scan_model_dir(model_dir), *scan_model_dir(extra_dir)]
+    *dirs: str | os.PathLike[str], huggingface: bool = True
+) -> OrderedDict[str, str]:
+    model_paths = []
+
+    for dir_ in dirs:
+        if not dir_:
+            continue
+        model_paths.extend(scan_model_dir(Path(dir_)))
 
     models = OrderedDict()
     if huggingface:
@@ -56,17 +69,17 @@ def get_models(
                 "hand_yolov8n.pt": hf_download("hand_yolov8n.pt"),
                 "person_yolov8n-seg.pt": hf_download("person_yolov8n-seg.pt"),
                 "person_yolov8s-seg.pt": hf_download("person_yolov8s-seg.pt"),
-                "yolov8x-world.pt": hf_download(
-                    "yolov8x-world.pt", repo_id="Bingsu/yolo-world-mirror"
+                "yolov8x-worldv2.pt": hf_download(
+                    "yolov8x-worldv2.pt", repo_id="Bingsu/yolo-world-mirror"
                 ),
             }
         )
     models.update(
         {
-            "mediapipe_face_full": None,
-            "mediapipe_face_short": None,
-            "mediapipe_face_mesh": None,
-            "mediapipe_face_mesh_eyes_only": None,
+            "mediapipe_face_full": "mediapipe_face_full",
+            "mediapipe_face_short": "mediapipe_face_short",
+            "mediapipe_face_mesh": "mediapipe_face_mesh",
+            "mediapipe_face_mesh_eyes_only": "mediapipe_face_mesh_eyes_only",
         }
     )
 
@@ -133,3 +146,11 @@ def create_bbox_from_mask(
         if bbox is not None:
             bboxes.append(list(bbox))
     return bboxes
+
+
+def ensure_pil_image(image: Any, mode: str = "RGB") -> Image.Image:
+    if not isinstance(image, Image.Image):
+        image = to_pil_image(image)
+    if image.mode != mode:
+        image = image.convert(mode)
+    return image