+++ tests/test_ultralytics.py
@@ -11,8 +11,10 @@ from adetailer.ultralytics import ultralytics_predict
         "face_yolov8n.pt",
         "face_yolov8n_v2.pt",
         "face_yolov8s.pt",
+        "face_yolov9c.pt",
         "hand_yolov8n.pt",
         "hand_yolov8s.pt",
+        "hand_yolov9c.pt",
         "person_yolov8n-seg.pt",
         "person_yolov8s-seg.pt",
         "person_yolov8m-seg.pt",
@@ -26,7 +28,7 @@ def test_ultralytics_hf_models(sample_image: Image.Image, model_name: str):
 
 
 def test_yolo_world_default(sample_image: Image.Image):
-    model_path = hf_hub_download("Bingsu/yolo-world-mirror", "yolov8x-world.pt")
+    model_path = hf_hub_download("Bingsu/yolo-world-mirror", "yolov8x-worldv2.pt")
     result = ultralytics_predict(model_path, sample_image)
     assert result.preview is not None
 
@@ -43,6 +45,6 @@ def test_yolo_world_default(sample_image: Image.Image):
     ],
 )
 def test_yolo_world(sample_image2: Image.Image, klass: str):
-    model_path = hf_hub_download("Bingsu/yolo-world-mirror", "yolov8x-world.pt")
+    model_path = hf_hub_download("Bingsu/yolo-world-mirror", "yolov8x-worldv2.pt")
     result = ultralytics_predict(model_path, sample_image2, classes=klass)
     assert result.preview is not None